Extract storymap grid
